requirements.txt
streamlit
streamlit-webrtc
opencv-python-headless
torch
torchvision
numpy
pandas
psutil
pillow
av
